requirements.txt
social-auth-app-django
django-star-ratings
